backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
gspread
python-dotenv
pypdf
google-generativeai